requirements.txt
flask
openai
python-dotenv
streamlit
numpy
huggingface-hub
sentence-transformers
PyMuPDF
lancedb